+++ packages/perf-js/src/RegExp.ts
@@ -1,9 +1,0 @@
-import { BenchmarkGroup } from '@easy/benchmark-js';
-
-const benchmark = new BenchmarkGroup();
-
-benchmark.add('RegExp#test', () => {
-    /o/.test('Hello World!');
-});
-
-benchmark.run();