+++ packages/benchmark-js/CHANGELOG.md
@@ -1,9 +1,13 @@
 # CHANGELOG
-## @easy/benchmark-js v0.3.0 (Unreleased)
+## @easy/benchmark-js v0.3.0 (2022-02-01)
 ### Features
 
-- Run pre-benchmarking double times for optimization.
 - Improve the accuracy of the minimum duration time.
+- Add method `BenchmarkGroup.setEmptyTest` to get the maximum ops as a base line.
+- Add methods `Benchmark.writeTesterCode` and `BenchmarkGroup.writeTestersCode` to print the tester source code.
+- Add support to pass arguments to `testFn`.
+- Add adjust-benchmarking stage after pre-benchmarking stage but before formal-benchmarking stage.
+- Add support to set selections for each `testFn`'s argument.
 
 ### Bug Fixes